% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{day01}
\alias{day01}
\alias{f01a_find_total_distance}
\alias{f01b_compute_similarity}
\alias{example_data_01}
\title{Day 01: Historian Hysteria}
\usage{
f01a_find_total_distance(x)

f01b_compute_similarity(x)

example_data_01(example = 1)
}
\arguments{
\item{x}{puzzle input}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f01a_find_total_distance(x)} returns the total
distance between the sorted ids. For Part Two,
\code{f01b_compute_similarity(x)} returns the total similarity score.
}
\description{
\href{https://adventofcode.com/2024/day/1}{Historian Hysteria}
}
\details{
\strong{Part One}

The \emph{Chief Historian} is always present for the big Christmas sleigh
launch, but nobody has seen him in months! Last anyone heard, he was
visiting locations that are historically significant to the North Pole;
a group of Senior Historians has asked you to accompany them as they
check the places they think he was most likely to visit.

As each location is checked, they will mark it on their list with a
\emph{star}. They figure the Chief Historian \emph{must} be in one of the first
fifty places they\'ll look, so in order to save Christmas, you need to
help them get \emph{fifty stars} on their list before Santa takes off on
December 25th.

Collect stars by solving puzzles. Two puzzles will be made available on
each day in the Advent calendar; the second puzzle is unlocked when you
complete the first. Each puzzle grants \emph{one star}. Good luck!

You haven\'t even left yet and the group of Elvish Senior Historians has
already hit a problem: their list of locations to check is currently
\emph{empty}. Eventually, someone decides that the best place to check first
would be the Chief Historian\'s office.

Upon pouring into the office, everyone confirms that the Chief Historian
is indeed nowhere to be found. Instead, the Elves discover an assortment
of notes and lists of historically significant locations! This seems to
be the planning the Chief Historian was doing before he left. Perhaps
these notes can be used to determine which locations to search?

Throughout the Chief\'s office, the historically significant locations
are listed not by name but by a unique number called the \emph{location ID}.
To make sure they don\'t miss anything, The Historians split into two
groups, each searching the office and trying to create their own
complete list of location IDs.

There\'s just one problem: by holding the two lists up \emph{side by side}
(your puzzle input), it quickly becomes clear that the lists aren\'t
very similar. Maybe you can help The Historians reconcile their lists?

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{3   4
4   3
2   5
1   3
3   9
3   3
}\if{html}{\out{</div>}}

Maybe the lists are only off by a small amount! To find out, pair up the
numbers and measure how far apart they are. Pair up the \emph{smallest number
in the left list} with the \emph{smallest number in the right list}, then the
\emph{second-smallest left number} with the \emph{second-smallest right number},
and so on.

Within each pair, figure out \emph{how far apart} the two numbers are;
you\'ll need to \emph{add up all of those distances}. For example, if you
pair up a \code{3} from the left list with a \code{7} from the right list, the
distance apart is \code{4}; if you pair up a \code{9} with a \code{3}, the distance
apart is \code{6}.

In the example list above, the pairs and distances would be as follows:
\itemize{
\item The smallest number in the left list is \code{1}, and the smallest number
in the right list is \code{3}. The distance between them is \emph{\code{2}}.
\item The second-smallest number in the left list is \code{2}, and the
second-smallest number in the right list is another \code{3}. The
distance between them is \emph{\code{1}}.
\item The third-smallest number in both lists is \code{3}, so the distance
between them is \emph{\code{0}}.
\item The next numbers to pair up are \code{3} and \code{4}, a distance of \emph{\code{1}}.
\item The fifth-smallest numbers in each list are \code{3} and \code{5}, a distance
of \emph{\code{2}}.
\item Finally, the largest number in the left list is \code{4}, while the
largest number in the right list is \code{9}; these are a distance \emph{\code{5}}
apart.
}

To find the \emph{total distance} between the left list and the right list,
add up the distances between all of the pairs you found. In the example
above, this is \code{2 + 1 + 0 + 1 + 2 + 5}, a total distance of \emph{\code{11}}!

Your actual left and right lists contain many location IDs. \emph{What is the
total distance between your lists?}

\strong{Part Two}

Your analysis only confirmed what everyone feared: the two lists of
location IDs are indeed very different.

Or are they?

The Historians can\'t agree on which group made the mistakes \emph{or} how to
read most of the Chief\'s handwriting, but in the commotion you notice
an interesting detail: a
lot
of location IDs appear in both lists! Maybe the other numbers aren\'t
location IDs at all but rather misinterpreted handwriting.

This time, you\'ll need to figure out exactly how often each number from
the left list appears in the right list. Calculate a total \emph{similarity
score} by adding up each number in the left list after multiplying it by
the number of times that number appears in the right list.

Here are the same example lists again:

\if{html}{\out{<div class="sourceCode">}}\preformatted{3   4
4   3
2   5
1   3
3   9
3   3
}\if{html}{\out{</div>}}

For these example lists, here is the process of finding the similarity
score:
\itemize{
\item The first number in the left list is \code{3}. It appears in the right
list three times, so the similarity score increases by
\verb{3 * 3 = }\emph{\code{9}}.
\item The second number in the left list is \code{4}. It appears in the right
list once, so the similarity score increases by \verb{4 * 1 = }\emph{\code{4}}.
\item The third number in the left list is \code{2}. It does not appear in the
right list, so the similarity score does not increase (\code{2 * 0 = 0}).
\item The fourth number, \code{1}, also does not appear in the right list.
\item The fifth number, \code{3}, appears in the right list three times; the
similarity score increases by \emph{\code{9}}.
\item The last number, \code{3}, appears in the right list three times; the
similarity score again increases by \emph{\code{9}}.
}

So, for these example lists, the similarity score at the end of this
process is \emph{\code{31}} (\code{9 + 4 + 0 + 0 + 9 + 9}).

Once again consider your left and right lists. \emph{What is their similarity
score?}
}
\examples{
f01a_find_total_distance(example_data_01())
f01b_compute_similarity(example_data_01())
}
