% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{day02}
\alias{day02}
\alias{f02a_count_safe_reports}
\alias{f02b_count_safe_dampened_reports}
\alias{example_data_02}
\title{Day 02: Red-Nosed Reports}
\usage{
f02a_count_safe_reports(x)

f02b_count_safe_dampened_reports(x)

example_data_02(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f02a_count_safe_reports(x)} returns the number of safe
reports. For Part Two, \code{f02b_count_safe_dampened_reports(x)} returns the
number of safe reports if we discard one number from each report.
}
\description{
\href{https://adventofcode.com/2024/day/2}{Red-Nosed Reports}
}
\details{
\strong{Part One}

Check whether lists of numbers are strictly increasing and differences less
than 3.

\strong{Part Two}

Check how many lists satisfy the above requirement if one element can
be dropped/ignored from a list.
}
\examples{
f02a_count_safe_reports(example_data_02())
f02b_count_safe_dampened_reports(example_data_02())
}
