% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03}
\alias{day03}
\alias{f03a_compute_muls}
\alias{f03b_compute_conditional_muls}
\alias{example_data_03}
\title{Day 03: Mull It Over}
\usage{
f03a_compute_muls(x)

f03b_compute_conditional_muls(x)

example_data_03(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f03a_compute_muls(x)} returns the sum of mul()
statements. For Part Two, \code{f03b_compute_conditional_muls(x)} returns the
sum of mul() statements.
}
\description{
\href{https://adventofcode.com/2024/day/3}{Mull It Over}
}
\details{
\strong{Part One}

Extract mul(x,y) instructions, run them, sum them.

\strong{Part Two}

Extract mul(x,y) instructions, run them, sum them, but ignore any mul()s
between don't() and do() statements.
}
\examples{
f03a_compute_muls(example_data_03())
f03b_compute_conditional_muls(example_data_03(2))
}
