% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04}
\alias{day04}
\alias{f04a_count_xmas}
\alias{f04b_count_mas_x}
\alias{example_data_04}
\title{Day 04: Ceres Search}
\usage{
f04a_count_xmas(x)

f04b_count_mas_x(x)

example_data_04(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f04a_count_xmas(x)} returns the number of XMAS/SMAX in
the word search. For Part Two, \code{f04b_count_mas_x(x)} returns the number of
MAS/SAM xs in the grid.
}
\description{
\href{https://adventofcode.com/2024/day/4}{Ceres Search}
}
\details{
\strong{Part One}

Find the word XMAS in a word search.

\strong{Part Two}

Find cases in the word search where two copies of MAS form an X.
}
\examples{
f04a_count_xmas(example_data_04())
f04b_count_mas_x(example_data_04())
}
