% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05}
\alias{day05}
\alias{f05a_check_sequences_and_sum_midpoints}
\alias{f05b_repair_sequences_and_sum_midpoints}
\alias{example_data_05}
\title{Day 05: Print Queue}
\usage{
f05a_check_sequences_and_sum_midpoints(x)

f05b_repair_sequences_and_sum_midpoints(x)

example_data_05(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f05a_check_sequences_and_sum_midpoints(x)} returns the
sums of middle element in each well-ordered list. For Part Two, \code{f05b(x)}
returns fixes the lists that are not well-ordered and sums the middle
elements.
}
\description{
\href{https://adventofcode.com/2024/day/5}{Print Queue}
}
\details{
\strong{Part One}

Check that lists of numbers (second half of input) are well-ordered according
to some rules (first half). Rules have form "X|Y" which means that X must
precede Y in an orderings with both. Return the sum of the middle number in
each well ordered list.

\strong{Part Two}

Find a repair for the incorrectly-ordered lists and compute the sum.
}
\examples{
f05a_check_sequences_and_sum_midpoints(example_data_05())
f05b_repair_sequences_and_sum_midpoints(example_data_05())
}
