% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{day07}
\alias{day07}
\alias{f07a_add_mul}
\alias{f07b_add_mul_concat}
\alias{example_data_07}
\title{Day 07: Bridge Repair}
\usage{
f07a_add_mul(x)

f07b_add_mul_concat(x)

example_data_07(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f07a_add_mul(x)} returns sum of elements that can
reconstructed with multiplication or addition. For Part Two,
\code{f07b_add_mul_concat(x)} returns sum of elements that can reconstructed
with multiplication, addition or concatenation.
}
\description{
\href{https://adventofcode.com/2024/day/7}{Bridge Repair}
}
\details{
\strong{Part One}

Given a list of patterns like \verb{X: A B C}, determine whether there is a way to
compute \code{X} through some combination of adding or multiplying elements in
\code{A},\code{B},\code{C}. In other words, solve whether \code{X == A ? B ? C} where \verb{?} can be
\code{+} or \code{*}. Precedence is strictly left to right. Return sum of X values in
list that can be reconstructed in this way.

\strong{Part Two}

\verb{?} can also now be concatenation. Find the same sum.
}
\examples{
f07a_add_mul(example_data_07())
f07b_add_mul_concat(example_data_07())
}
