% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day09.R
\name{day09}
\alias{day09}
\alias{f09a_compute_checksum}
\alias{f09b_reorder_and_compute_checksum}
\alias{example_data_09}
\title{Day 09: Disk Fragmenter}
\usage{
f09a_compute_checksum(x)

f09b_reorder_and_compute_checksum(x)

example_data_09(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f09a_compute_checksum(x)} returns the checksum. For
Part Two, \code{f09b_reorder_and_compute_checksum(x)} returns the checksum.
}
\description{
\href{https://adventofcode.com/2024/day/9}{Disk Fragmenter}
}
\details{
A sequence of digits like \code{12345} is a disk map where each digit indicates
the size of a region on the disk. The map alternates between files and free
spaces. Each file is uniquely IDed by its position. So \code{12345} expands into
\verb{0..111....22222}: a file of size 1, free space of size 2, etc.

\strong{Part One}

Compact the disk by moving individual pieces of memory from the right end to
the leftmost free space. A compacted \code{12345} map yields \verb{022111222......}.
Multiply each 0-based position by its ID value and take the sum.

\strong{Part Two}

Compact the disk moving whole files by decreasing ID into the leftmost
region of contiguous free space that can fit the file. Multiply
each 0-based position by its ID value and take the sum.
}
\examples{
f09a_compute_checksum(example_data_09())
f09b_reorder_and_compute_checksum(example_data_09())
}
