% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{day10}
\alias{day10}
\alias{f10a_score_trailheads}
\alias{f10b_rate_trailheads}
\alias{example_data_10}
\title{Day 10: Hoof It}
\usage{
f10a_score_trailheads(x)

f10b_rate_trailheads(x)

example_data_10(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f10a_score_trailheads(x)} returns the solution. For
Part Two, \code{f10b_rate_trailheads(x)} returns the solution.
}
\description{
\href{https://adventofcode.com/2024/day/10}{Hoof It}
}
\details{
Given a topographic map (grid of numbers), a path on the map is a path from 0
to 1 to 2 to ... 9. A path can only move in orthogonal directions (up, left,
right, down).

\strong{Part One}

Find the total number of 9s that are reachable from each 0. Sum the counts.

\strong{Part Two}

Find the total number of unique paths from each 0 to all reachable 9s. Sum
the counts.
}
\examples{
f10a_score_trailheads(example_data_10(4))
f10b_rate_trailheads(example_data_10(4))
}
