% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{day01}
\alias{day01}
\alias{f01a_find_total_distance}
\alias{f01b_compute_similarity}
\alias{example_data_01}
\title{Day 01: Historian Hysteria}
\usage{
f01a_find_total_distance(x)

f01b_compute_similarity(x)

example_data_01(example = 1)
}
\arguments{
\item{x}{puzzle input}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f01a_find_total_distance(x)} returns the total
distance between the sorted ids. For Part Two,
\code{f01b_compute_similarity(x)} returns the total similarity score.
}
\description{
\href{https://adventofcode.com/2024/day/1}{Historian Hysteria}
}
\details{
\strong{Part One}

Sort two columns, find total absolute difference in each row.

\strong{Part Two}

Multiple each number in one column by its frequency in another column. Sum
these products.
}
\examples{
f01a_find_total_distance(example_data_01())
f01b_compute_similarity(example_data_01())
}
