% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{day02}
\alias{day02}
\alias{f02a_count_safe_reports}
\alias{f02b_count_safe_dampened_reports}
\alias{example_data_02}
\title{Day 02: Red-Nosed Reports}
\usage{
f02a_count_safe_reports(x)

f02b_count_safe_dampened_reports(x)

example_data_02(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f02a_count_safe_reports(x)} returns the number of safe
reports. For Part Two, \code{f02b_count_safe_dampened_reports(x)} returns the
number of safe reports if we discard one number from each report.
}
\description{
\href{https://adventofcode.com/2024/day/2}{Red-Nosed Reports}
}
\details{
\strong{Part One}

Fortunately, the first location The Historians want to search isn\'t a
long walk from the Chief Historian\'s office.

While the \href{/2015/day/19}{Red-Nosed Reindeer nuclear fusion/fission plant} appears to contain no sign of the Chief Historian,
the engineers there run up to you as soon as they see you. Apparently,
they \emph{still} talk about the time Rudolph was saved through molecular
synthesis from a single electron.

They\'re quick to add that - since you\'re already here - they\'d really
appreciate your help analyzing some unusual data from the Red-Nosed
reactor. You turn to check if The Historians are waiting for you, but
they seem to have already divided into groups that are currently
searching every corner of the facility. You offer to help with the
unusual data.

The unusual data (your puzzle input) consists of many \emph{reports}, one
report per line. Each report is a list of numbers called \emph{levels} that
are separated by spaces. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{7 6 4 2 1
1 2 7 8 9
9 7 6 2 1
1 3 2 4 5
8 6 4 4 1
1 3 6 7 9
}\if{html}{\out{</div>}}

This example data contains six reports each containing five levels.

The engineers are trying to figure out which reports are \emph{safe}. The
Red-Nosed reactor safety systems can only tolerate levels that are
either gradually increasing or gradually decreasing. So, a report only
counts as safe if both of the following are true:
\itemize{
\item The levels are either \emph{all increasing} or \emph{all decreasing}.
\item Any two adjacent levels differ by \emph{at least one} and \emph{at most
three}.
}

In the example above, the reports can be found safe or unsafe by
checking those rules:
\itemize{
\item \verb{7 6 4 2 1}: \emph{Safe} because the levels are all decreasing by 1 or 2.
\item \verb{1 2 7 8 9}: \emph{Unsafe} because \verb{2 7} is an increase of 5.
\item \verb{9 7 6 2 1}: \emph{Unsafe} because \verb{6 2} is a decrease of 4.
\item \verb{1 3 2 4 5}: \emph{Unsafe} because \verb{1 3} is increasing but \verb{3 2} is
decreasing.
\item \verb{8 6 4 4 1}: \emph{Unsafe} because \verb{4 4} is neither an increase or a
decrease.
\item \verb{1 3 6 7 9}: \emph{Safe} because the levels are all increasing by 1, 2,
or 3.
}

So, in this example, \emph{\code{2}} reports are \emph{safe}.

Analyze the unusual data from the engineers. \emph{How many reports are
safe?}

\strong{Part Two}

The engineers are surprised by the low number of safe reports until they
realize they forgot to tell you about the Problem
Dampener.

The Problem Dampener is a reactor-mounted module that lets the reactor
safety systems \emph{tolerate a single bad level} in what would otherwise be
a safe report. It\'s like the bad level never happened!

Now, the same rules apply as before, except if removing a single level
from an unsafe report would make it safe, the report instead counts as
safe.

More of the above example\'s reports are now safe:
\itemize{
\item \verb{7 6 4 2 1}: \emph{Safe} without removing any level.
\item \verb{1 2 7 8 9}: \emph{Unsafe} regardless of which level is removed.
\item \verb{9 7 6 2 1}: \emph{Unsafe} regardless of which level is removed.
\item \code{1 }\emph{\code{3}}\verb{ 2 4 5}: \emph{Safe} by removing the second level, \code{3}.
\item \verb{8 6 }\emph{\code{4}}\verb{ 4 1}: \emph{Safe} by removing the third level, \code{4}.
\item \verb{1 3 6 7 9}: \emph{Safe} without removing any level.
}

Thanks to the Problem Dampener, \emph{\code{4}} reports are actually \emph{safe}!

Update your analysis by handling situations where the Problem Dampener
can remove a single level from unsafe reports. \emph{How many reports are now
safe?}
}
\examples{
f02a_count_safe_reports(example_data_02())
f02b_count_safe_dampened_reports(example_data_02())
}
